requirements.txt
# SynthForge.AI - Azure Architecture Diagram Analyzer
# Dependencies for Microsoft Agent Framework-based implementation
# NO STATIC MAPPINGS - Uses official Microsoft Azure Architecture Icons

# Azure AI Agents SDK (Microsoft Agent Framework)
azure-ai-agents>=1.0.0
azure-ai-projects>=1.0.0

# Azure SDK Core
azure-identity>=1.15.0
azure-core>=1.30.0

# Azure AI Vision - Image Analysis 4.0 (RECOMMENDED for diagram OCR)
# Uses DefaultAzureCredential - no API key needed!
# Best for scene text extraction from architecture diagrams
azure-ai-vision-imageanalysis>=1.0.0

# Azure Computer Vision - Legacy SDK (fallback if above not available)
# Requires explicit AZURE_VISION_KEY configuration
# azure-cognitiveservices-vision-computervision>=0.9.0
# msrest>=0.7.0

# Azure Document Intelligence (alternative OCR for documents)
# Better for structured documents like forms/invoices
# azure-ai-documentintelligence>=1.0.0

# Pydantic for data models
pydantic>=2.5.0

# Image handling
pillow>=10.0.0
opencv-python>=4.8.0  # For icon template/feature matching

# SVG to PNG conversion for enhanced icon matching
# Requires Cairo graphics library on your system:
#   Windows: Install GTK3 runtime or use conda
#   macOS: brew install cairo
#   Linux: sudo apt install libcairo2-dev
# cairosvg>=2.7.0

# CLI enhancements (optional but recommended)
rich>=13.0.0

# Environment management
python-dotenv>=1.0.0

# YAML for configuration
pyyaml>=6.0

# HTTP client for MCP and icon download
httpx>=0.25.0
aiohttp>=3.9.0

# Type hints
typing-extensions>=4.8.0
